Add Vue
We'll be using vue as our frontent framework; in the long run we will
want to set up TypeScript and either a bundler or something that serves
the content with hot-loading etc.  For now we're just going to write the
code manually in rapid-fire R&D mode, which is why we're including the
source directly instead of via node + npm.

Issue #7