Add proposal seed
The sample proposal data is now loaded into the database on DB seed.

Our DB types are a bit of a disaster right now, and the schema overall
should be considered unsafe for real data, as we'll be dropping the
tables with real types (and real migrations) at some point in the future
if this project exits demo mode.

Issue #12