Replace db setup with peewee
This commit breaks the API completely, but is a necessary transition
point.

The logic for creating and seeding the database has been moved away from
raw SQL and over to using the ORM.

The db creation is no longer done in the context of the application;
this may not be the "right" approach but I could not find much direciton
either way.  We may want to consider moving to the FlaskDB helper object
that is provided by peewee.

Issue #25